feat: traverosors need to use template for default base conversion context
maybe we can get rid of base class at all to use only smart pointer of a templated type, only thing that matters is to have merge function
