+++ vrml_proc/src/traversors/handlers/BoxHandler.hpp
@@ -6,11 +6,14 @@
 #include "BaseConversionContextActionMap.hpp"
 #include "FullParsedVrmlNodeContext.hpp"
 
+#include "VrmlProcessingExport.hpp"
+
 namespace vrml_proc {
 	namespace traversor {
 		namespace handler {
 			namespace BoxHandler {
-				std::shared_ptr<vrml_proc::conversion_context::BaseConversionContext> Handle(vrml_proc::traversor::FullParsedVrmlNodeContext context, const vrml_proc::action::BaseConversionContextActionMap& actionMap);
+				template<typename DefaultConversionContext>
+				VRMLPROCESSING_API std::shared_ptr<vrml_proc::conversion_context::BaseConversionContext> Handle(vrml_proc::traversor::FullParsedVrmlNodeContext context, const vrml_proc::action::BaseConversionContextActionMap& actionMap);
 			}
 		}
 	}
